libgif.js: optimized by replacing reduce with for-loop and bitwise operators
I didn't measure the performance changes as a direct result of this commit but reduce is generally slower.
I really wanted a drastic improvement in performance for my application and this should help a little.
It looked like the modified function was called frequently.

Some general discussion of reduce performance is at:
https://stackoverflow.com/questions/43556132/javascript-performance-reduce-vs-for-loop

Some performance tests results were done at:
https://leanylabs.com/blog/js-forEach-map-reduce-vs-for-for_of/